Change type of rating in ratings tables to unsignedTinyInteger

+++ server/database/migrations/2019_11_13_175444_create_poll_ratings_table.php
@@ -20,7 +20,7 @@ class CreatePollRatingsTable extends Migration
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->unsignedBigInteger('item_id');
             $table->foreign('item_id')->references('id')->on('items')->onDelete('cascade');
-            $table->integer('rating')->nullable();
+            $table->unsignedTinyInteger('rating')->nullable();
             $table->timestamps();
         });
     }
